Move model-related things from Tenancy to a trait
